added mutations to type def and renamed model folders name. Started to change resolvers
