+++ server/models/index.js
@@ -1,6 +1,6 @@
 const User = require('./User');
 const Product = require('./Product');
-const Category = require('./Category');
+const Category = require('./Planet');
 const Order = require('./Order');
 
 module.exports = { User, Product, Category, Order };
